Corrects what it doesn't need JavaScript for
I could turn the html into pug so I can put the queue in on the back end. So you can get the queue by refreshing. Because the way it's going to turn out right now it wouuld collect it with JavaScript.

But who am I kidding, I don't want to bother with that.

I think I prefer HTML syntax to pug anyway.

Sorry Christopher.