Removes binaries
I never thought they were a good idea, because runners of the program would have to trust the binaries given, and I didn't want to saddle comp soc with dependecies on their machine.
I only did it because music.get does it, and thought that must be the method they prefer, but then again mpv wasn't included as a binary in music.get, so it should all be possible.