Broadcast the queue when a new song starts
Still dealing with the issues of what order things are loaded in and when functions become available.

I've drawn up a graph irl. It's almost a tree but not quite. I think there's some stuff to change. I don't think main needs to be the inbetween for wsServer any more.

Content Manager especially I don't think needs to require main more than once. But it looks like http server might because ws is defined after... hold on for the next commit.