Changes Credits
Adds the WordArt guy.
Puts the ClippyJS forker in the same format as the others
Moves the accredition for 2 of them to where the code is in the repo because although I used and modified their solutions, putting them here suggests that I couldn't have done those things by myself, when really I found them while looking for an existing solution built into nodejs.