Reduce complexity of queue code
Now currPosteriority is only set to a user when their item is chosen. So now if 2 people have not played anything in a long time, the one who's waited the longest goes next. But they are immediately put to the currPosteriority which will be boosted when the item finishes playing.